% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse.R
\name{engines}
\alias{engines}
\title{Language engines}
\usage{
engines(...)
}
\arguments{
\item{...}{Named values (for setting) or unnamed values (for getting).}
}
\description{
Get or set language engines for evaluating code chunks and inline code.
}
\details{
The usage is similar to \code{\link[=reactor]{reactor()}}, e.g., \code{engines('LANG')} returns an
engine function for the language \code{LANG}, and \code{engines(LANG = function(x, inline = FALSE, ...) {})} sets the engine for a language.

An engine function should have three arguments:
\itemize{
\item \code{x}: An element in the \code{\link[=crack]{crack()}} list (a code chunk or a text block).
\item \code{inline}: It indicates if \code{x} is from a code chunk or inline code.
\item \code{...}: Currently unused but recommended for future compatibility (more
arguments might be passed to the function).
}

The function should return a character value.
}
\examples{
litedown::engines()  # built-in engines
}
