% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse.R
\name{fuse_exit}
\alias{fuse_exit}
\title{Exit \code{fuse()} early}
\usage{
fuse_exit(append = "")
}
\arguments{
\item{append}{A character string to be appended to the fuse output after the
current code chunk.}
}
\value{
Invisible \code{NULL}.
}
\description{
Stop \code{\link[=fuse]{fuse()}} from processing the remaining code chunks and text blocks. This
function is meant to be called inside a code chunk to end the document early.
}
