% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse.R
\name{get_input}
\alias{get_input}
\title{Get the input file path of \code{fuse()}}
\usage{
get_input()
}
\value{
A character string, or \code{NULL} if the input is not a file (e.g., when
using the \code{text} argument of \code{fuse()}).
}
\description{
A helper function to return the input file path of \code{\link[=fuse]{fuse()}} when called
inside a code chunk.
}
