% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pkg_desc}
\alias{pkg_desc}
\alias{pkg_news}
\alias{pkg_citation}
\alias{pkg_manual}
\title{Get the package description, news, citation, and manual pages}
\usage{
pkg_desc(name = detect_pkg())

pkg_news(name = detect_pkg(), path = detect_news(name), recent = 1, ...)

pkg_citation(name = detect_pkg())

pkg_manual(name = detect_pkg())
}
\arguments{
\item{name}{The package name (by default, it is automatically detected from
the \code{DESCRIPTION} file if it exists in the current working directory or
upper-level directories).}

\item{path}{Path to the \code{NEWS.md} file. If empty, \code{\link[=news]{news()}} will be called to
retrieve the news entries.}

\item{recent}{The number of recent versions to show. By default, only the
latest version's news entries are retrieved. To show the full news, set
\code{recent = 0}.}

\item{...}{Other arguments to be passed to \code{\link[=news]{news()}}.}
}
\value{
A character vector (HTML or Markdown) that will be printed as is
inside a code chunk of an R Markdown document.

\code{pkg_desc()} returns an HTML table containing the package metadata.

\code{pkg_news()} returns the news entries.

\code{pkg_citation()} returns the package citation in both the plain-text
and BibTeX formats.

\code{pkg_manual()} returns all manual pages of the package in HTML.
}
\description{
Helper functions to retrieve various types of package information that can be
put together as the full package documentation like a \pkg{pkgdown} website.
These functions can be called inside any R Markdown document.
}
\examples{
litedown::pkg_desc()
litedown::pkg_news()
litedown::pkg_citation()
litedown::pkg_manual()
}
