% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{roam}
\alias{roam}
\title{Preview Markdown and R Markdown files}
\usage{
roam(dir = ".", live = TRUE, ...)
}
\arguments{
\item{dir}{A directory path.}

\item{live}{Whether to enable live preview. If enabled, the browser page will
be automatically updated upon modification of local files used by the page
(e.g., the Markdown file or external CSS/JS/image files). If disabled, you
can manually refresh the page to fully re-render it.}

\item{...}{Other arguments to be passed to \code{\link[xfun:new_app]{xfun::new_app()}}.}
}
\description{
Launch a web page to list and preview files under a directory.
}
\details{
Markdown files will be converted to HTML and returned to the web browser
directly without writing to HTML files, to keep the directory clean during
the preview. Clicking on a filename will bring up an HTML preview. To see its
raw content, click on the link on its file size instead.
}
