% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{smartypants}
\alias{smartypants}
\title{Convert some ASCII strings to HTML entities}
\usage{
smartypants(text)
}
\arguments{
\item{text}{A character vector of the Markdown text.}
}
\value{
A character vector of the transformed text.
}
\description{
Transform ASCII strings \code{(c)} (copyright), \code{(r)} (registered trademark),
\code{(tm)} (trademark), and fractions \code{n/m} into \emph{smart} typographic HTML
entities.
}
\examples{
cat(litedown:::smartypants("1/2 (c)\n"))
}
