% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{html_format}
\alias{html_format}
\alias{latex_format}
\title{Output formats in YAML metadata}
\usage{
html_format(options = NULL, meta = NULL, template = NULL, keep_md = FALSE)

latex_format(
  options = NULL,
  meta = NULL,
  template = NULL,
  keep_md = FALSE,
  keep_tex = FALSE,
  latex_engine = "xelatex"
)
}
\arguments{
\item{meta, options}{Arguments to be passed to \code{\link[=mark]{mark()}}.}

\item{template}{A template file path.}

\item{keep_md, keep_tex}{Whether to keep the intermediate \file{.md} and
\file{.tex} files generated from \file{.Rmd}.}

\item{latex_engine}{The LaTeX engine to compile \file{.tex} to \file{.pdf}.}
}
\description{
The primary output formats of \pkg{litedown} are HTML and LaTeX. These output
formats can be configured in the \code{output} field of the YAML metadata of the
Markdown document.
}
\details{
The \pkg{rmarkdown} package converts Markdown using Pandoc by default, and it
also accept custom converting tools. The output formats here provide the
custom converting function \code{\link[=mark]{mark()}} to \pkg{rmarkdown}, so that users can
take advantage of \code{\link[rmarkdown:render]{rmarkdown::render()}} and the Knit button in RStudio. It is
absolutely not necessary to rely on \pkg{rmarkdown}. The only point is
convenience. If you do not use \code{rmarkdown::render()} or the Knit button, you
can definitely just call \code{litedown::mark()} directly.
}
