% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{litedown-package}
\alias{litedown}
\alias{litedown-package}
\title{A lightweight version of R Markdown}
\description{
Markdown is a plain-text format that can be converted to HTML and other
formats. This package can render R Markdown to Markdown, and then to an
output document format. The main differences between this package and
\pkg{rmarkdown} are that it does not use Pandoc or \pkg{knitr} (i.e., fewer
dependencies), and it also has fewer Markdown features.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yihui/litedown}
  \item Report bugs at \url{https://github.com/yihui/litedown/issues}
}

}
\author{
\strong{Maintainer}: Yihui Xie \email{xie@yihui.name} (\href{https://orcid.org/0000-0003-0645-5666}{ORCID}) (URL: https://yihui.org)

Authors:
\itemize{
  \item Yihui Xie \email{xie@yihui.name} (\href{https://orcid.org/0000-0003-0645-5666}{ORCID}) (URL: https://yihui.org)
}

Other contributors:
\itemize{
  \item Tim Taylor (\href{https://orcid.org/0000-0002-8587-7113}{ORCID}) [contributor]
  \item John MacFarlane (Author of cmark) [copyright holder]
  \item Jeroen Ooms (Author of the R bindings to cmark in the commonmark package, from which the C glue code was adapted) [contributor, copyright holder]
  \item GitHub, Inc. (Author of the cmark-gfm extensions) [copyright holder]
}

}
