% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{markdown_options}
\alias{markdown_options}
\title{Markdown rendering options}
\usage{
markdown_options()
}
\value{
A character vector of all available options.
}
\description{
A list of all options to control Markdown rendering. Options that are enabled
by default are marked by a \code{+} prefix, and those disabled by default are
marked by \code{-}.
}
\details{
See \url{https://yihui.org/litedown/#sec:markdown-options} for the full list of
options and their documentation.
}
\examples{
# all available options
litedown::markdown_options()
}
